Add template selection drop-down box.
The original assignment assumed one had access to these files on their
hard drive. This might not be the case for an employer browsing the
portfolio.

The example files are to be bundled up in a *.zip that can be browsed
once the web page is fully loaded. Selecting a file will load the
grammar from the zip and parse it.